// src/app/admin/admin-shared/admin.interfaces.ts
export interface Admin {
  ACCESS_TOKEN: string;
  ACCT_NO: string;
  ACTIVE: number;
  ADDRESS: string;
  AZURE_ID: string;
  BUSINESS_NAME: string;
  CITY: string;
  CONFIRM_TOKEN: string;
  CONFIRM_TOKEN_EXPIRES: string;
  CONFIRM_TOKEN_TYPE: string;
  CREATED_AT: string;
  EMAIL: string;
  EMAIL_VERIFIED: number;
  EMAIL_VERIFIED_TIME: string;
  FAX: string;
  FIRST_NAME: string;
  LANG: string;
  LAST_NAME: string;
  LOGIN_ATTEMPTS: number;
  PHONE: string;
  POSTAL_CODE: string;
  PROVINCE: string;
  REFRESH_TOKEN: string;
  ROLE: number;
  USER_ID: number;
}